perf: dont process global search selectors when not searching (#6298)
* chore: break out search results from GlobalSearchButton

* chore: break out search modal from GlobalSearchButton
